+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/openai
 
+## 4.0.0-beta.9
+
+### Patch Changes
+
+- 156cdf0: feat(openai): add new tool search tool
+
+## 4.0.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [f7d4f01]
+  - @ai-sdk/provider-utils@5.0.0-beta.3
+  - @ai-sdk/provider@4.0.0-beta.2
+
+## 4.0.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [5c2a5a2]
+  - @ai-sdk/provider@4.0.0-beta.1
+  - @ai-sdk/provider-utils@5.0.0-beta.2
+
 ## 4.0.0-beta.6
 
 ### Patch Changes