+++ packages/react/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ai-sdk/react
 
+## 4.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [34fd051]
+  - ai@7.0.0-beta.21
+
+## 4.0.0-beta.20
+
+### Patch Changes
+
+- ai@7.0.0-beta.20
+
+## 4.0.0-beta.19
+
+### Patch Changes
+
+- Updated dependencies [f7d4f01]
+  - @ai-sdk/provider-utils@5.0.0-beta.3
+  - ai@7.0.0-beta.19
+
+## 4.0.0-beta.18
+
+### Patch Changes
+
+- ai@7.0.0-beta.18
+- @ai-sdk/provider-utils@5.0.0-beta.2
+
 ## 4.0.0-beta.17
 
 ### Patch Changes